+++ lib/wavefront-cli/commands/serviceaccount.rb
@@ -2,31 +2,37 @@
 
 require_relative 'base'
 
-# Define the user command.
+# Define the service account command.
 #
-class WavefrontCommandUser < WavefrontCommandBase
-  def description
-    "view and manage Wavefront #{things}"
+class WavefrontCommandServiceaccount < WavefrontCommandBase
+  def thing
+    'service account'
   end
 
-  # delete uses a different string because it accepts multiples.
-  # Adding ellipsis to <id> causes everything else to expect an
-  # array.
-  #
+  def sdk_file
+    'serviceaccount'
+  end
+
+  def sdk_class
+    'ServiceAccount'
+  end
+
   def _commands
     ["list #{CMN} [-l] [-O fields]",
      "describe #{CMN} <id>",
-     "create #{CMN} [-e] [-m permission...] [-g group...] <id>",
-     "invite #{CMN} [-m permission...] [-g group...] <id>",
-     "delete #{CMN} <user>...",
+     "create #{CMN} [-I] [-d description] [-m permission...] [-g group...] " \
+     '[-k token...] <id>',
+     "activate #{CMN} <id>",
+     "deactivate #{CMN} <id>",
      "dump #{CMN}",
-     "import #{CMN} [-u] <file>",
      "groups #{CMN} <id>",
      "privileges #{CMN} <id>",
      "join #{CMN} <id> <group>...",
      "leave #{CMN} <id> <group>...",
      "grant #{CMN} <privilege> to <id>",
      "revoke #{CMN} <privilege> from <id>",
+     "set #{CMN} <key=value> <id>",
+     "import #{CMN} [-u] <file>",
      "search #{CMN} [-al] [-o offset] [-L limit] <condition>..."]
   end
 
@@ -38,14 +44,10 @@ class WavefrontCommandUser < WavefrontCommandBase
      "-L, --limit=COUNT        number of #{things} to list",
      '-O, --fields=F1,F2,...   only show given fields',
      "-u, --update             update an existing #{thing}",
-     "-e, --email              send e-mail to #{thing} on account creation",
+     "-I, --inactive           create an inactive #{thing}",
+     "-d, --desc=STRING        description of #{thing}",
      "-m, --permission=STRING  give #{thing} this permission",
-     "-g, --group=STRING       add #{thing} to this user group"]
-  end
-
-  def postscript
-    'If your account does not have RBAC enabled, you must grant user ' \
-    "permissions with '-m'. For a list of permissions, see " \
-    "'wf usergroup --help'.".fold(TW, 0)
+     "-g, --group=STRING       add #{thing} to this user group",
+     '-k, --apitoken=STRING    API token']
   end
 end