+++ spec/test_mixins/dump.rb
@@ -16,11 +16,12 @@ module WavefrontCliTest
       out, err = capture_io do
         assert_raises(SystemExit) do
           assert_cmd_gets('dump --format=json',
-                          "/api/v2/#{api_class}?limit=999&offset=0")
+                          "/api/v2/#{api_path}?limit=999&offset=0")
         end
         Spy.teardown
       end
 
+      assert_cannot_noop('dump --format=json')
       assert_empty(err)
       assert_equal([].to_s, out.rstrip)
     end
@@ -29,11 +30,12 @@ module WavefrontCliTest
       out, err = capture_io do
         assert_raises(SystemExit) do
           assert_cmd_gets('dump --format=yaml',
-                          "/api/v2/#{api_class}?limit=999&offset=0")
+                          "/api/v2/#{api_path}?limit=999&offset=0")
         end
         Spy.teardown
       end
 
+      assert_cannot_noop('dump --format=yaml')
       assert_empty(err)
       assert_equal('--- []', out.rstrip)
     end