+++ HISTORY.md
@@ -1,5 +1,9 @@
 # Changelog
 
+* Add `serviceaccount` command.
+* Add `user privileges` command.
+* Require 3.4.x of [the SDK](https://github.com/snltd/wavefront-sdk).
+
 ## 4.2.2
 * Refactor of all tests, which exposed the following:
 * Fix noops on `proxy versions`, `window ongoing` and `window